+++ src/ranbval_sdk/exceptions/__init__.py
@@ -36,8 +36,8 @@ re-exported here, so ``from ranbval_sdk.exceptions import RanbvalConfigError`` i
 from ranbval_sdk.exceptions.base import RanbvalError
 from ranbval_sdk.exceptions.config import MissingKeyError, RanbvalConfigError
 from ranbval_sdk.exceptions.crypto import RanbvalDecryptError, RanbvalSecurityError
-from ranbval_sdk.exceptions.policy import RepoNotAllowedError, RepoPolicyError
 from ranbval_sdk.exceptions.plan import PlanLimitError
+from ranbval_sdk.exceptions.policy import RepoNotAllowedError, RepoPolicyError
 from ranbval_sdk.exceptions.proxy import ProxyError
 
 __all__ = [